Convert hooks useCreateNoticeForMultipleErrors and useCreateNoticeForRejectedPromises into standard functions
